+++ sdk/core/azure_core/src/process/tokio.rs
@@ -1,21 +1,21 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-use super::ProcessRunner;
+use super::Executor;
 use async_trait::async_trait;
-use std::io::Error;
-use std::{ffi::OsStr, process::Output};
+use std::{ffi::OsStr, io, process::Output};
 
-/// Implement [`ProcessRunner`] via [`::tokio::process`]
+/// An [`Executor`] using [`tokio::process::Command`].
 #[derive(Debug)]
-pub struct TokioRunner;
+pub struct TokioExecutor;
 
 #[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
 #[cfg_attr(not(target_arch = "wasm32"), async_trait)]
-impl ProcessRunner for TokioRunner {
-    async fn run_command(&self, program: &OsStr, args: &[&OsStr]) -> Result<Output, Error> {
-        let mut cmd = ::tokio::process::Command::new(program);
-        cmd.args(args);
-        cmd.output().await
+impl Executor for TokioExecutor {
+    async fn run(&self, program: &OsStr, args: &[&OsStr]) -> io::Result<Output> {
+        ::tokio::process::Command::new(program)
+            .args(args)
+            .output()
+            .await
     }
 }