+++ sdk/identity/azure_identity/examples/azure_cli_credentials.rs
@@ -8,20 +8,21 @@ use url::Url;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn Error>> {
-    let sub_id = AzureCliCredential::get_subscription().await?;
-    println!("Azure cli subscription: {sub_id}");
+    let subscription_id =
+        std::env::var("AZURE_SUBSCRIPTION_ID").expect("AZURE_SUBSCRIPTION_ID required");
 
-    let credentials = AzureCliCredential::new()?;
+    let credentials = AzureCliCredential::new(None)?;
     let res = credentials
         .get_token(&["https://management.azure.com/.default"])
         .await?;
-    println!("Azure cli response == {res:?}");
+    eprintln!("Azure CLI response == {res:?}");
+
     // Let's enumerate the Azure storage accounts
     // in the subscription. Note: this way of calling the REST API
     // will be different (and easier) using other Azure Rust SDK
     // crates, this is just an example.
     let url = Url::parse(&format!(
-                 "https://management.azure.com/subscriptions/{sub_id}/providers/Microsoft.Storage/storageAccounts?api-version=2019-06-01"
+                 "https://management.azure.com/subscriptions/{subscription_id}/providers/Microsoft.Storage/storageAccounts?api-version=2019-06-01"
              ))?;
 
     let resp = reqwest::Client::new()