+++ sdk/identity/azure_identity/src/lib.rs
@@ -37,13 +37,64 @@ fn test_validate_not_empty() {
     assert!(validate_not_empty("not empty", "it's not empty").is_ok());
 }
 
+#[cfg_attr(target_arch = "wasm32", allow(dead_code))]
+fn validate_scope(scope: &str) -> Result<()> {
+    if scope.is_empty()
+        || !scope.chars().all(|c| {
+            c.is_alphanumeric() || c == '.' || c == '-' || c == '_' || c == ':' || c == '/'
+        })
+    {
+        return Err(Error::message(
+            ErrorKind::Credential,
+            format!("invalid scope {scope}"),
+        ));
+    }
+
+    Ok(())
+}
+
+#[test]
+fn test_validate_scope() {
+    assert!(validate_scope("").is_err());
+    assert!(validate_scope("invalid_scope@id").is_err());
+    assert!(validate_scope("A-1b_2c 3:d/4.z").is_ok());
+    assert!(validate_scope("http://vault.azure.net").is_ok());
+}
+
+#[cfg_attr(target_arch = "wasm32", allow(dead_code))]
+fn validate_subscription(subscription: &str) -> Result<()> {
+    if subscription.is_empty()
+        || !subscription
+            .chars()
+            .all(|c| c.is_alphabetic() || c == '.' || c == '-' || c == '_' || c == ' ')
+    {
+        return Err(Error::message(
+            ErrorKind::Credential,
+            format!("invalid subscription {subscription}. If this is the name of a subscription, use its ID instead"),
+        ));
+    }
+
+    Ok(())
+}
+
+#[test]
+fn test_validate_subscription() {
+    assert!(validate_subscription("").is_err());
+    assert!(validate_subscription("invalid_subscription@id").is_err());
+    assert!(validate_subscription("A-1b_2c 3.z").is_ok());
+    assert!(validate_subscription("7b795fb9-09d3-42f4-a494-38864f99ba3c").is_ok());
+}
+
 fn validate_tenant_id(tenant_id: &str) -> Result<()> {
     if tenant_id.is_empty()
         || !tenant_id
             .chars()
             .all(|c| c.is_alphanumeric() || c == '.' || c == '-')
     {
-        return Err(Error::message(ErrorKind::Credential, "invalid tenantID. You can locate your tenantID by following the instructions listed here: https://learn.microsoft.com/partner-center/find-ids-and-domain-names"));
+        return Err(Error::message(
+            ErrorKind::Credential,
+            format!("invalid tenant ID {tenant_id}. You can locate your tenantID by following the instructions listed here: https://learn.microsoft.com/partner-center/find-ids-and-domain-names"),
+        ));
     }
 
     Ok(())